adds WithConverter() to SubscriberOption
