+++ projector_test.go
@@ -2,6 +2,7 @@ package gomessagestore_test
 
 import (
 	"context"
+	"io/ioutil"
 	"testing"
 
 	. "github.com/blackhatbrigade/gomessagestore"
@@ -27,6 +28,7 @@ func TestProjectorAcceptsAReducer(t *testing.T) {
 	mockRepo := mock_repository.NewMockRepository(ctrl)
 
 	var logrusLogger = logrus.New()
+	logrusLogger.Out = ioutil.Discard
 	myMessageStore := NewMessageStoreFromRepository(mockRepo, logrusLogger)
 
 	mockred := new(mockReducer1)
@@ -52,6 +54,7 @@ func TestProjectorAcceptsAReducerFunc(t *testing.T) {
 	mockRepo := mock_repository.NewMockRepository(ctrl)
 
 	var logrusLogger = logrus.New()
+	logrusLogger.Out = ioutil.Discard
 	myMessageStore := NewMessageStoreFromRepository(mockRepo, logrusLogger)
 
 	mockred := func(msg Message, previousState interface{}) interface{} {
@@ -85,6 +88,7 @@ func TestProjectorAcceptsADefaultState(t *testing.T) {
 	mockRepo := mock_repository.NewMockRepository(ctrl)
 
 	var logrusLogger = logrus.New()
+	logrusLogger.Out = ioutil.Discard
 	myMessageStore := NewMessageStoreFromRepository(mockRepo, logrusLogger)
 
 	defstate := mockDataStructure{}
@@ -112,6 +116,7 @@ func TestProjectorRunsWithReducers(t *testing.T) {
 	mockRepo := mock_repository.NewMockRepository(ctrl)
 
 	var logrusLogger = logrus.New()
+	logrusLogger.Out = ioutil.Discard
 	myMessageStore := NewMessageStoreFromRepository(mockRepo, logrusLogger)
 
 	mockReducerFunc := func(msg Message, previousState interface{}) interface{} {
@@ -186,6 +191,7 @@ func TestProjectorPicksUpAfterFullBatch(t *testing.T) {
 	mockRepo := mock_repository.NewMockRepository(ctrl)
 
 	var logrusLogger = logrus.New()
+	logrusLogger.Out = ioutil.Discard
 	myMessageStore := NewMessageStoreFromRepository(mockRepo, logrusLogger)
 
 	defstate := mockDataStructure{}
@@ -256,6 +262,7 @@ func TestCreateProjectorFailsIfGivenPointerForDefaultState(t *testing.T) {
 	mockRepo := mock_repository.NewMockRepository(ctrl)
 
 	var logrusLogger = logrus.New()
+	logrusLogger.Out = ioutil.Discard
 	myMessageStore := NewMessageStoreFromRepository(mockRepo, logrusLogger)
 
 	defstate := new(mockDataStructure)
@@ -278,6 +285,7 @@ func TestCreateProjectorFailsIfDefaultStateIsNotSet(t *testing.T) {
 	mockRepo := mock_repository.NewMockRepository(ctrl)
 
 	var logrusLogger = logrus.New()
+	logrusLogger.Out = ioutil.Discard
 	myMessageStore := NewMessageStoreFromRepository(mockRepo, logrusLogger)
 
 	_, err := myMessageStore.CreateProjector(
@@ -298,6 +306,7 @@ func TestCreateProjectorFailsWithoutAtLeastOneReducer(t *testing.T) {
 	mockRepo := mock_repository.NewMockRepository(ctrl)
 
 	var logrusLogger = logrus.New()
+	logrusLogger.Out = ioutil.Discard
 	myMessageStore := NewMessageStoreFromRepository(mockRepo, logrusLogger)
 
 	defstate := mockDataStructure{}
